+++ Magic/CHANGELOG.md
@@ -66,6 +66,7 @@
  - Add `self_destruct` wand property, for a wand that will be destroyed when held
  - Spell parameters can now be used inside of equations within that spell
  - The `ModifyAttribute` action now supports an `x` variable holding the current attribute value, just like `ModifyProperties`
+ - The `/mage deactivate` command can now be used as an alias for `/mage lock`
 
 ## Changes